Add fs::create_directory() and move fs utility function to their own file
